pyproject.toml
[tool.poetry]
name = "Argus"
version = "0.1.0"
description = "A staged static analysis tool for finding security vulnerabilites in workflows"
authors = ["Siddharth Muralee <siddharth.muralee@gmail.com>"]
license = "GPL-3"

[tool.poetry.dependencies]
python = "^3.8"
pathlib = "^1.0.1"
click = "^8.1.3"
GitPython = "^3.1.31"
colorlog = "^6.7.0"
PyYAML = "^6.0"
sarif-om = "^1.0.4"
jschema-to-python = "^1.2.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
